# docs/02-api-specification.yaml
openapi: 3.0.3
info:
  title: CMS kéo-thả API
  version: 0.2.0
  description: >
    API cho Editor Engine, Content Editor UI và Renderer.
    Mọi endpoint yêu cầu header Authorization: Bearer <token> trừ khi ghi chú khác.
    Phiên bản 0.2.0 — Cập nhật sau SA Review 17/08/2026:
    Bổ sung pagination, refresh token, DELETE/PATCH, siteId trên content-items, Error schema chuẩn, Sites_Users management, media upload per-site, X-Request-ID.
servers:
  - url: https://api.example.com/v1
...
paths:
...
  /sites/{siteId}/members:
    get:
      tags: [Sites]
      summary: Danh sách thành viên trong site (chỉ owner/designer)
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Danh sách thành viên
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SiteUser' }
    post:
      tags: [Sites]
      summary: Mời thành viên mới vào site (chỉ owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string, format: email }
                role:  { type: string, enum: [designer, editor, viewer] }
      responses:
        '201':
          description: Đã mời thành công
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SiteUser' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /sites/{siteId}/members/{userId}:
    patch:
      tags: [Sites]
      summary: Đổi role thành viên (chỉ owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [designer, editor, viewer] }
      responses:
        '200':
          description: Đã cập nhật
    delete:
      tags: [Sites]
      summary: Xóa thành viên khỏi site (chỉ owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Đã xóa

  # ─── PAGES ───────────────────────────────────────────────────
  /sites/{siteId}/pages:
    get:
      tags: [Pages]
      summary: Danh sách trang trong 1 site
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - name: status
          in: query
          schema: { type: string, enum: [draft, published] }
      responses:
        '200':
          description: Danh sách page (paginated)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedPages' }
    post:
      tags: [Pages]
      summary: Tạo page mới (chỉ role designer/owner)
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug]
              properties:
                slug:     { type: string }
                seo_meta: { type: object, additionalProperties: true }
      responses:
        '201':
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /sites/{siteId}/pages/{pageId}:
    get:
      tags: [Pages]
      summary: Lấy thông tin 1 page
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
    patch:
      tags: [Pages]
      summary: Cập nhật page (slug, seo_meta)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                slug:     { type: string }
                seo_meta: { type: object, additionalProperties: true }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Page' }
    delete:
      tags: [Pages]
      summary: Xóa mềm page (chỉ owner/designer)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
      responses:
        '204':
          description: Đã xóa (soft delete)
  /sites/{siteId}/pages/{pageId}/versions:
    get:
      tags: [Pages]
      summary: Lịch sử version của 1 page
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedPageVersions' }
    post:
      tags: [Pages]
      summary: Tạo version mới (autosave / submit review)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [draft, in_review] }
                blocks:
                  type: array
                  items: { $ref: '#/components/schemas/Block' }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageVersion' }
  /sites/{siteId}/pages/{pageId}/versions/{versionId}/approve:
    post:
      tags: [Pages]
      summary: Duyệt version (chỉ owner, và không phải người đã submit)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
        - $ref: '#/components/parameters/VersionId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                review_notes: { type: string, description: "Ghi chú duyệt" }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageVersion' }
        '403':
          $ref: '#/components/responses/Forbidden'
  /sites/{siteId}/pages/{pageId}/versions/{versionId}/publish:
    post:
      tags: [Pages]
      summary: Publish version đã approved — kích hoạt Renderer revalidate
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/PageId'
        - $ref: '#/components/parameters/VersionId'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  published_url: { type: string, format: uri }

  # ─── BLOCKS ──────────────────────────────────────────────────
...
  /sites/{siteId}/media:
    get:
      tags: [Media]
      summary: Danh sách media assets của site (paginated)
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagedMediaAssets' }
    post:
      tags: [Media]
      summary: Khởi tạo upload — backend trả presigned PUT URL, client upload trực tiếp lên MinIO
      parameters:
        - $ref: '#/components/parameters/SiteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, mime_type, file_size]
              properties:
                filename:  { type: string }
                mime_type: { type: string, example: "image/jpeg" }
                file_size: { type: integer, description: "bytes" }
      responses:
        '200':
          description: Presigned URL để client upload trực tiếp
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_url: { type: string, format: uri, description: "PUT trực tiếp lên MinIO" }
                  asset_id:   { type: string, format: uuid, description: "ID để confirm sau khi upload" }
        '413':
          description: File vượt quota của site
  /sites/{siteId}/media/{assetId}/confirm:
    post:
      tags: [Media]
      summary: Xác nhận upload thành công — backend lưu bản ghi vào media_assets
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/AssetId'
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaAsset' }
  /sites/{siteId}/media/{assetId}:
    delete:
      tags: [Media]
      summary: Xóa mềm media asset
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - $ref: '#/components/parameters/AssetId'
      responses:
        '204':
          description: Đã xóa (soft delete; file MinIO dọn bởi cleanup job)

  # ─── AI ──────────────────────────────────────────────────────
...
components:
  parameters:
    SiteId:
      name: siteId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageId:
      name: pageId
      in: path
      required: true
      schema: { type: string, format: uuid }
    VersionId:
      name: versionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    AssetId:
      name: assetId
      in: path
      required: true
      schema: { type: string, format: uuid }
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
    Cursor:
      name: cursor
      in: query
      description: "Cursor-based pagination — giá trị từ trường `next_cursor` của response trước"
      schema: { type: string }
  responses:
...
    Forbidden:
      description: Không đủ quyền
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:       { type: string, example: "RESOURCE_NOT_FOUND", description: "Machine-readable error code" }
        message:    { type: string, example: "Page không tồn tại", description: "Human-readable message" }
        details:    { type: object, additionalProperties: true, description: "Context bổ sung (field validation errors...)" }
        request_id: { type: string, description: "X-Request-ID để trace log" }
    Pagination:
      type: object
      properties:
        next_cursor: { type: string, nullable: true, description: "null nếu hết trang" }
        has_more:    { type: boolean }
        total:       { type: integer, description: "Tổng số records (optional, tính riêng nếu cần)" }
...
    SiteUser:
      type: object
      properties:
        user_id:    { type: string, format: uuid }
        email:      { type: string }
        name:       { type: string }
        role:       { type: string, enum: [owner, designer, editor, viewer] }
        created_at: { type: string, format: date-time }
    Page:
      type: object
      properties:
        id:                   { type: string, format: uuid }
        site_id:              { type: string, format: uuid }
        current_version_id:   { type: string, format: uuid, nullable: true }
        slug:                 { type: string }
        status:               { type: string, enum: [draft, published] }
        seo_meta:             { type: object, additionalProperties: true }
        created_at:           { type: string, format: date-time }
        updated_at:           { type: string, format: date-time }
    PagedPages:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Page' }
    PageVersion:
      type: object
      properties:
        id:             { type: string, format: uuid }
        page_id:        { type: string, format: uuid }
        version_number: { type: integer }
        status:         { type: string, enum: [draft, in_review, approved, published] }
        submitted_by:   { type: string, format: uuid, nullable: true }
        approved_by:    { type: string, format: uuid, nullable: true }
        review_notes:   { type: object, additionalProperties: true, nullable: true }
        created_at:     { type: string, format: date-time }
    PagedPageVersions:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/PageVersion' }
    Block:
      type: object
      properties:
        id:               { type: string, format: uuid }
        page_version_id:  { type: string, format: uuid }
        parent_id:        { type: string, format: uuid, nullable: true }
        type:             { type: string, example: heading }
        props:            { type: object, additionalProperties: true }
        order_index:      { type: integer }
...
    MediaAsset:
      type: object
      properties:
        id:         { type: string, format: uuid }
        site_id:    { type: string, format: uuid }
        url:        { type: string, format: uri }
        type:       { type: string, enum: [image, video] }
        mime_type:  { type: string }
        file_size:  { type: integer, description: "bytes" }
        filename:   { type: string }
        created_at: { type: string, format: date-time }
    PagedMediaAssets:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/MediaAsset' }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...